Modifique errores de compilacion en todas estas clases. Ademàs modifique el Makefile para que compile a source/client/common-ui

+++ trunk/source/client/common-ui/uisupport.cpp
@@ -2,10 +2,10 @@
 
 bool UISupport::isMouseOver(const UIMouseInformation& mouseInformation, int x,
     int y, int width, int height) {
-    if ( (mouseInformation.getX() < x) ||
-         (mouseInformation.getY() < y) ||
-         (mouseInformation.getX() >= x + width) ||
-         (mouseInformation.getY() >= y + height) )
+    if ( (((int) mouseInformation.getX()) < x) ||
+         (((int) mouseInformation.getY()) < y) ||
+         (((int) mouseInformation.getX()) >= x + width) ||
+         (((int) mouseInformation.getY()) >= y + height) )
         return false;
     return true;
 }
